Converted build info API to un/pack functions. Converted packstr function to send string size as uint16_t (used to 	be uint32_t, which wasted space).

+++ trunk/src/common/pack.h
@@ -23,9 +23,9 @@ void	_unpack32(uint32_t *valp, void **bufp, int *lenp);
 void	_pack16(uint16_t val, void **bufp, int *lenp);
 void	_unpack16(uint16_t *valp, void **bufp, int *lenp);
 
-void	_packstr(char *valp, uint32_t size_val, void **bufp, int *lenp);
-void	_unpackstr_ptr(char **valp, uint32_t *size_valp, void **bufp, int *lenp);
-void	_unpackstr_xmalloc(char **valp, uint32_t *size_valp, void **bufp, int *lenp);
+void	_packstr(char *valp, uint16_t size_val, void **bufp, int *lenp);
+void	_unpackstr_ptr(char **valp, uint16_t *size_valp, void **bufp, int *lenp);
+void	_unpackstr_xmalloc(char **valp, uint16_t *size_valp, void **bufp, int *lenp);
 
 #define pack32(val,bufp,lenp) do {			\
 	assert(sizeof(val) == sizeof(uint32_t)); 	\
@@ -63,7 +63,7 @@ void	_unpackstr_xmalloc(char **valp, uint32_t *size_valp, void **bufp, int *lenp
 
 #define packstr(valp,size_val,bufp,lenp) do {		\
 	assert(size_val == 0 || valp != NULL);		\
-	assert(sizeof(size_val) == sizeof(uint32_t));	\
+	assert(sizeof(size_val) == sizeof(uint16_t));	\
 	assert((bufp) != NULL && *(bufp) != NULL);	\
         assert((lenp) != NULL);				\
         assert(*(lenp) >= (sizeof(size_val)+size_val));	\
@@ -72,7 +72,7 @@ void	_unpackstr_xmalloc(char **valp, uint32_t *size_valp, void **bufp, int *lenp
 
 #define unpackstr_ptr(valp,size_valp,bufp,lenp) do {	\
 	assert(valp != NULL);				\
-	assert(sizeof(size_valp) == sizeof(uint32_t *));\
+	assert(sizeof(size_valp) == sizeof(uint16_t *));\
 	assert((bufp) != NULL && *(bufp) != NULL);	\
         assert((lenp) != NULL);				\
         assert(*(lenp) >= sizeof(uint32_t));		\
@@ -81,7 +81,7 @@ void	_unpackstr_xmalloc(char **valp, uint32_t *size_valp, void **bufp, int *lenp
 
 #define unpackstr_xmalloc(valp,size_valp,bufp,lenp) do {\
 	assert(valp != NULL);				\
-	assert(sizeof(size_valp) == sizeof(uint32_t *));\
+	assert(sizeof(size_valp) == sizeof(uint16_t *));\
 	assert((bufp) != NULL && *(bufp) != NULL);	\
         assert((lenp) != NULL);				\
         assert(*(lenp) >= sizeof(uint32_t));		\
